+++ CHANGELOG.md
@@ -6,6 +6,11 @@ x.y.z Release notes (yyyy-MM-dd)
 <!-- ### Fixes - Include, if needed -->
 <!-- ### Internal - Include, if needed -->
 
+### Enhancements
+
+* Implement `isRemoteConfigAvailable` PWA communication handler to check if the remote config feature is generally available and returns `false` by default. This behavior can be overridden in `AppCallback.isRemoteConfigAvailable(isAvailable)`.
+
+
 10.1.1 Release notes (2021-09-30)
 =============================================================